refactor: move DomainError to shared
