+++ src/domain/ledger/errors.ts
@@ -1,4 +1,4 @@
-import { DomainError, ErrorLevel } from "@domain/errors"
+import { DomainError, ErrorLevel } from "@domain/shared"
 
 export class LedgerError extends DomainError {}
