+++ src/domain/shared/index.ts
@@ -1,2 +1,3 @@
 export * from "./primitives"
 export * from "./calculator"
+export * from "./errors"
