+++ src/domain/shared/index.types.d.ts
@@ -1,3 +1,6 @@
+type ErrorLevel =
+  typeof import("./errors").ErrorLevel[keyof typeof import("./errors").ErrorLevel]
+
 type WalletCurrency =
   typeof import("./primitives").WalletCurrency[keyof typeof import("./primitives").WalletCurrency]
